consolidate db.links module
